Split extension manager responsibilities
